// lib/api/generated/models/resultStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * CKS Slavia API
 * API panelu klubowego CKS Slavia Ruda Śląska
 * OpenAPI spec version: 0.1.0
 */

export type ResultStatus = typeof ResultStatus[keyof typeof ResultStatus];


export const ResultStatus = {
  pending: 'pending',
  accepted: 'accepted',
  rejected: 'rejected',
  needs_edit: 'needs_edit',
} as const;